Actions don't do anything for inputs.
Actions are shit for things like this and it's best to take it locally, i do not care, nor endorse, nor give a shit if you are scared of local development, i love it and stuff, besides, locally is so much better, and besides, are you lazy to type a single command in there? or is it something else that you don't want to tell me? Please open op an issue or go to #issues on the discord server or go to discussions and open/start one, now thank you for your time to read this!